# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SMS ANGOLA",
        "version": "1.0"
    },
    "paths": {
...
        "/api/phonebooks?num=1&limit=10&search=''": {
            "get": {
                "tags": [
                    "Phonebooks"
                ],
                "description": "Liste des contacts.",
                "operationId": "listContact",
                "responses": {
                    "200": {
                        "description": "Liste des contacts."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/phonebooks/{uid}": {
            "get": {
                "tags": [
                    "Phonebooks"
                ],
                "description": "Détail d'un contact.",
                "operationId": "showContact",
                "responses": {
                    "200": {
                        "description": "Détail d'un Contact."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Phonebooks"
                ],
                "description": "Modification d'un contact.",
                "operationId": "editContact",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "number"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "number"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "date_at": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "field1": {
                                        "type": "string"
                                    },
                                    "field2": {
                                        "type": "string"
                                    },
                                    "field3": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact modifié avec succès."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/api/publipostage?num=1&limit=10&search=''": {
            "get": {
                "tags": [
                    "Publipostage"
                ],
                "description": "Liste des contacts.",
                "operationId": "listPub",
                "responses": {
                    "200": {
                        "description": "Liste des contacts."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/publipostage/{uid}": {
            "get": {
                "tags": [
                    "Publipostage"
                ],
                "description": "Détail d'un contact.",
                "operationId": "showPub",
                "responses": {
                    "200": {
                        "description": "Détail d'un Contact."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Publipostage"
                ],
                "description": "Modification d'un contact.",
                "operationId": "editPub",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "number"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "number"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "date_at": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "field1": {
                                        "type": "string"
                                    },
                                    "field2": {
                                        "type": "string"
                                    },
                                    "field3": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact modifié avec succès."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/publipostage": {
            "post": {
                "tags": [
                    "Publipostage"
                ],
                "description": "Enregistrement d'un contact.",
                "operationId": "storePub",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "label",
                                    "number"
                                ],
                                "properties": {
                                    "label": {
                                        "type": "string"
                                    },
                                    "number": {
                                        "type": "number"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "date_at": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "field1": {
                                        "type": "string"
                                    },
                                    "field2": {
                                        "type": "string"
                                    },
                                    "field3": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact enregisté avec succès."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/publipostage/delete": {
            "delete": {
                "tags": [
                    "Publipostage"
                ],
                "description": "Suppression d'un ou de plusieurs contacts.",
                "operationId": "deletePub",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "contacts"
                                ],
                                "properties": {
                                    "contacts": {
                                        "type": "array",
                                        "items": {
                                            "example": "['335d5855-31b1-44f7-81fd-56e7e7c82a07', 'e504f670-a605-4827-a148-77746018e83f', 'd77b6ee9-8121-4d84-9678-290f7988248d']"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contacts supprimés avec succès."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/publipostage/imports": {
            "post": {
                "tags": [
                    "Publipostage"
                ],
                "description": "Importation d'un contact.",
                "operationId": "importPub",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "files"
                                ],
                                "properties": {
                                    "files": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact enregisté avec succès."
                    },
                    "400": {
                        "description": "Serveur indisponible."
                    },
                    "404": {
                        "description": "Page introuvable."
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}